base_script v3.3

+++ install/sh/repo_one.sh
@@ -12,4 +12,14 @@ echo "...
 Скрипт по замене репозитория на РедОС выполнен."
 echo "
 "
-reboot
+clear
+echo -n "
+Перезагрузить ПК?(да/нет): "
+
+read askReboot
+if [ $askReboot = "да" ]
+then
+    reboot
+else 
+    ./install.sh
+fi
